requirements.txt
dvc[gdrive]==2.41.1
ffmpeg-python==0.2.0
hydra-core==1.3.1
imutils==0.5.4
joblib==1.2.0
matplotlib==3.6.2
numpy==1.24.1
omegaconf==2.3.0
opencv-python==4.7.0.68
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
pre-commit==2.21.0
pydicom==2.3.1
pytest==7.2.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
supervisely==6.69.2
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
segmentation_models_pytorch==0.3.2
pytorch_lightning==2.0.0
tensorboard-2.12.1

